Creates an active dbplyr tbl connection to a database

open_tbl_con() takes a list of database and schema arguments and connects to a table in the database schema using the function tbl.

Since this function returns an active connection to the database it is up to the user to close the connection when they are done. It is recommended to not use open_tbl_con() directly, but instead use a (project-specific) wrapper function that ensures the connection gets closed once the data is accessed (see examples).

Usage

open_tbl_con(con_specs)

Arguments

con_specs

List. Connection arguments for the database. It should be a named list with the names "name"/"con_args"/"tbl_args". con_args should be a named character with entries "server"/"database"/"uid"/"pwd" and tbl_args should be a named character with the entries "dbcatalog"(optional)/dbschema"/"dbtable".

Value

A list containing an active tbl and connection. Remember to close the connection once tbl has been used.

Examples

if (FALSE) { # \dontrun{
# import the above yaml
passwords <- read_yaml('passwords.yaml')

# import function that wraps around open_tbl_con
# This could be placed in your projects functions/ folder
# or just on top of you import script to keep the import command
# easily accessible.
get_tab <- function(db_args){
  tbl_con <- open_tbl_con(db_args)

  # this ensures the connection always gets closed properly
  on.exit(dbDisconnect(tbl_con$con))

  # This part is project specific. filters and selects etc before
  # the collect() command will be done outside of R and may
  # significantly increase the fetch time
  # returns date of birth and municipality for 3 days in the data
  tbl_con$tbl |>
    filter(date >= as.Date("2023-01-01"),
           date  < as.Date("2023-01-04")) |>
    select(date_of_birth, municipality) |>
    collect()
}

# fetch data using wrapper function
tab <- get_tab(passwords$prod)

} # }